# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudOptix API
  version: "1.0.0"
  summary: >-
    Autonomous cloud cost intelligence and optimization: architecture-aware
    discovery, economics, recommendations, governed automation and an
    auditable execution trail, all tenant-isolated.
  description: |
    CloudOptix models a customer's AWS estate as a digital twin, prices it,
    finds savings, and — within policy — executes and rolls back changes
    itself. This document describes the whole HTTP API a client (the
    CloudOptix web app, a CI pipeline calling the cost compiler, or a partner
    integration) talks to.

    ## Authentication
    Every route except `/healthz`, `/readyz`, `/metrics`, `/openapi.yaml` and
    the `Onboarding` tag requires a credential: either an OIDC-issued bearer
    JWT (`Authorization: Bearer <token>`) or a service API key
    (`X-CloudOptix-Api-Key: <key>`). See the `BearerAuth` and `ApiKeyAuth`
    security schemes below. A worker or CI system uses a service token
    (issued out of band) presented the same way as a user's bearer token.

    ## Tenant scope
    An OIDC principal may belong to more than one tenant; which one a
    request is scoped to is resolved from the token's own tenant claim when
    present, or from the `X-CloudOptix-Tenant` header otherwise, with a
    membership check either way — a header can select among tenants a
    principal actually belongs to, never grant a new one. A service token
...
    ever names a tenant explicitly.

...

...

...

...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# YAML anchors for parameters and error responses shared across dozens of
# operations. A vendor extension at the document root (x-*, permitted
# anywhere per the OpenAPI 3.1 Specification Extensions rule) rather than a
# components/ entry, because these are resolved by the YAML parser itself —
# before any OpenAPI tooling sees the document — while a $ref into
# components/parameters would still repeat one line per operation for no
# benefit; the effect on the parsed document is identical to writing every
# field out longhand at every one of the ~115 operations below, just without
# ~115 chances for one of those copies to drift from the rest.
...
paths:
...
  /approvals:
    parameters: [*tenantHeader, *cursorParam, *limitParam, *sortParam, *orderParam]
    get:
      operationId: approvals.list
      tags: [Approvals]
      summary: List approval requests
      responses:
        "200":
          description: A page of approvals.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalPage" }
        <<: *commonErrors
    post:
      operationId: approvals.request
      tags: [Approvals]
      summary: Manually request an approval
      parameters: [*idempotencyKey]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalRequest" }
      responses:
        "201":
          description: Requested.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequest" }
        <<: *commonErrors
...
  /automation/process:
    parameters: [*tenantHeader, *idempotencyKey]
    post:
      operationId: automation.process
      tags: [Automation]
      summary: Run one autonomous optimization cycle
      description: >-
        Finds recommendations the active policy permits to auto-execute,
        plans them, executes them within the maintenance window and
        concurrency limits, and schedules validation. Normally invoked by a
        scheduled worker, exposed here for on-demand and support use.
      responses:
        "200":
          description: Cycle result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutonomousRunResult" }
        <<: *commonErrors
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        An OIDC-issued user JWT, an internal service token, or (in
        non-production deployments only) the fixed development static
        token. Validated per internal/infrastructure/auth: issuer, audience,
        expiry and signing algorithm are all checked, and `alg: none` is
        always rejected.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-CloudOptix-Api-Key
      description: A long-lived service API key, bound to exactly one tenant at issuance.
...
  schemas:
...
    RiskLevel:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
...
    Confidence:
      type: number
      minimum: 0
      maximum: 1
      description: A 0..1 confidence score.
    Money:
      type: object
      description: An exact monetary amount, stored as integer micros to avoid floating-point drift.
      required: [micros, currency, amount, display]
      properties:
        micros: { type: integer, format: int64, example: 4110000000, description: "Exact amount in 1/1,000,000ths of a unit." }
        currency: { type: string, example: USD }
        amount: { type: number, example: 4110.00, description: "Convenience float; micros is authoritative." }
        display: { type: string, example: "$4,110.00" }
...
    PageEnvelope:
      type: object
      description: The cursor-pagination envelope every list endpoint shares.
      properties:
        next_cursor: { type: string }
        total: { type: integer }

    # -------------------------------------------------------------
    # AWS Accounts
    # -------------------------------------------------------------
...
    ApprovalContext:
      type: object
      properties:
        monthly_saving: { $ref: "#/components/schemas/Money" }
        annual_saving: { $ref: "#/components/schemas/Money" }
        monthly_cost_delta: { $ref: "#/components/schemas/Money" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        risk_level: { $ref: "#/components/schemas/RiskLevel" }
        blast_summary: { type: string }
        environment: { type: string }
        affected_resources: { type: array, items: { type: string } }
        rollback_plan: { type: string }
        validation_plan: { type: string }
        policy_reason: { type: string }
        diff: { type: string }
    ApprovalResponse:
      type: object
      properties:
        principal: { type: string }
        role: { type: string }
        approved: { type: boolean }
        comment: { type: string }
        at: { type: string, format: date-time }
        ip_address: { type: string }
        user_agent: { type: string }
    ApprovalRequest:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        subject_kind: { type: string, enum: [execution_plan, policy, aws_connection, commitment_purchase] }
        subject_id: { type: string }
        title: { type: string }
        summary: { type: string }
        context: { $ref: "#/components/schemas/ApprovalContext" }
        policy_decision_id: { type: string }
        required_roles: { type: array, items: { type: string } }
        min_approvals: { type: integer }
        require_distinct_approver: { type: boolean }
        state: { type: string, enum: [pending, approved, rejected, expired, withdrawn] }
        responses: { type: array, items: { $ref: "#/components/schemas/ApprovalResponse" } }
        requested_by: { type: string }
        requested_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time }
        execute_after: { type: string, format: date-time }
    ApprovalPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/ApprovalRequest" } }
...
    AutonomousRunResult:
      type: object
      properties:
        considered: { type: integer }
        planned: { type: integer }
        executed: { type: integer }
        skipped: { type: integer }
        failed: { type: integer }
        rolled_back: { type: integer }
        monthly_saving: { $ref: "#/components/schemas/Money" }
        skip_reasons: { type: object, additionalProperties: { type: integer } }
        duration_ms: { type: integer }
...